third_party/WebKit/LayoutTests/fast/dom/inner-text-with-no-renderer.html

<!doctype>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
test(() => {
  const div = document.createElement('div');
  div.innerHTML = '&quot;Text&quot;<br>new line';
  assert_equals('"Text"\nnew line', div.innerText);
 },
'innerText for an element without layout object');
</script>
